refactor: consolidate command execution into workspace-profile
