+++ src/workspace-profile.test.ts
@@ -1,24 +1,23 @@
 import { describe, expect, test } from "bun:test";
-import { lintFiles } from "./lint-reflection";
 import type { WorkspaceCommand } from "./workspace-profile";
+import { runCommandWithFiles } from "./workspace-profile";
 
 const BIOME: WorkspaceCommand = { bin: "bunx", args: ["biome", "check"] };
 
-describe("lintFiles", () => {
+describe("runCommandWithFiles", () => {
   test("returns no errors for empty file list", () => {
-    const result = lintFiles("/tmp", [], BIOME);
+    const result = runCommandWithFiles("/tmp", BIOME, []);
     expect(result.hasErrors).toBe(false);
     expect(result.output).toBe("");
   });
 
   test("returns no errors for a clean file", () => {
-    const result = lintFiles(import.meta.dir, ["lint-reflection.ts"], BIOME);
+    const result = runCommandWithFiles(import.meta.dir, BIOME, ["workspace-profile.ts"]);
     expect(result.hasErrors).toBe(false);
   });
 
   test("treats errors on missing files as lint errors", () => {
-    const result = lintFiles(import.meta.dir, ["nonexistent-file.ts"], BIOME);
-    // biome exits non-zero for missing files — that's a valid lint error
+    const result = runCommandWithFiles(import.meta.dir, BIOME, ["nonexistent-file.ts"]);
     expect(result.hasErrors).toBe(true);
   });
 });
